Move GenericStreamAcceptor to a separate cpp file
Also added a UNIX socket usage example and  a big scary warning
concerning usage of external IP for the API  in config.lua.
