+++ src/CMakeLists.txt
@@ -5,6 +5,7 @@ set(tfs_SRC
 	${CMAKE_CURRENT_LIST_DIR}/http_api/responder.cpp
 	${CMAKE_CURRENT_LIST_DIR}/http_api/router.cpp
 	${CMAKE_CURRENT_LIST_DIR}/http_api/server.cpp
+	${CMAKE_CURRENT_LIST_DIR}/http_api/generic_stream_acceptor.cpp
 	${CMAKE_CURRENT_LIST_DIR}/ban.cpp
 	${CMAKE_CURRENT_LIST_DIR}/baseevents.cpp
 	${CMAKE_CURRENT_LIST_DIR}/bed.cpp
